202603 Zero linter warnings: isize + no panics (breaking) (#233)
- No panics. Instead, error propagation everywhere
- Switch to `isize` as the base type
- Breaking changes due to new signatures

+++ ailets-rs/actor_runtime/src/dagops.rs
@@ -19,7 +19,7 @@ use std::io::Write;
 /// # Errors
 ///
 /// - Normally, should never fail.
-pub fn value_node(value: &[u8], explain: &str) -> Result<i32, String> {
+pub fn value_node(value: &[u8], explain: &str) -> Result<isize, String> {
     let mut value_base64 = Vec::new();
     let mut enc =
         base64::write::EncoderWriter::new(&mut value_base64, &engine::general_purpose::STANDARD);
@@ -48,7 +48,7 @@ pub fn value_node(value: &[u8], explain: &str) -> Result<i32, String> {
 /// # Errors
 ///
 /// - Wrong handle.
-pub fn alias(alias: &str, node_handle: i32) -> Result<i32, String> {
+pub fn alias(alias: &str, node_handle: isize) -> Result<isize, String> {
     let alias = std::ffi::CString::new(alias).map_err(|e| e.to_string())?;
 
     let handle = unsafe { dag_alias(alias.as_ptr().cast::<i8>(), node_handle) };
@@ -95,8 +95,8 @@ pub fn detach_from_alias(alias: &str) -> Result<(), String> {
 /// - The host can fail
 pub fn instantiate_with_deps(
     workflow_name: &str,
-    deps: impl Iterator<Item = (String, i32)>,
-) -> Result<i32, String> {
+    deps: impl Iterator<Item = (String, isize)>,
+) -> Result<isize, String> {
     let workflow_name = std::ffi::CString::new(workflow_name).map_err(|e| e.to_string())?;
 
     let mut deps_json = serde_json::Map::new();
@@ -128,7 +128,7 @@ pub fn instantiate_with_deps(
 /// # Errors
 ///
 /// - Host runtime error
-pub fn open_write_pipe(explain: Option<&str>) -> Result<i32, String> {
+pub fn open_write_pipe(explain: Option<&str>) -> Result<isize, String> {
     let explain_cstr = if let Some(explain) = explain {
         Some(std::ffi::CString::new(explain).map_err(|e| e.to_string())?)
     } else {
@@ -158,7 +158,7 @@ pub fn open_write_pipe(explain: Option<&str>) -> Result<i32, String> {
 ///
 /// - Invalid file descriptor
 /// - Host runtime error
-pub fn alias_fd(alias: &str, fd: i32) -> Result<i32, String> {
+pub fn alias_fd(alias: &str, fd: isize) -> Result<isize, String> {
     let alias = std::ffi::CString::new(alias).map_err(|e| e.to_string())?;
 
     let handle = unsafe { dag_alias_fd(alias.as_ptr().cast::<i8>(), fd) };
